listener manager as a revealed function
