+++ src/utils/listenerMatch.js
@@ -1,0 +1,62 @@
+/*
+Copyright 2020 Adobe. All rights reserved.
+This file is licensed to you under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License. You may obtain a copy
+of the License at http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software distributed under
+the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+OF ANY KIND, either express or implied. See the License for the specific language
+governing permissions and limitations under the License.
+*/
+
+const constants = require('../constants');
+
+const has = require('lodash/has');
+
+/**
+ * Checks if the listener matches the item.
+ *
+ * @param {Listener} listener The listener.
+ * @param {Item} item The item.
+ * @returns {Boolean} true if listener matches the item, false otherwise.
+ * @private
+ */
+function listenerMatch(listener, item) {
+  const event = listener.event;
+  const itemConfig = item.config;
+  let matches = false;
+
+  if (item.type === constants.itemType.DATA) {
+    if (event === constants.dataLayerEvent.CHANGE) {
+      matches = selectorMatches(listener, item);
+    }
+  } else if (item.type === constants.itemType.EVENT) {
+    if (event === constants.dataLayerEvent.EVENT || event === itemConfig.event) {
+      matches = selectorMatches(listener, item);
+    }
+    if (item.data && event === constants.dataLayerEvent.CHANGE) {
+      matches = selectorMatches(listener, item);
+    }
+  }
+
+  return matches;
+}
+
+/**
+ * Checks if a listener has a selector that points to an object in the data payload of an item.
+ *
+ * @param {Listener} listener The listener to extract the selector from.
+ * @param {Item} item The item.
+ * @returns {Boolean} true if a selector is not provided or if the given selector is matching, false otherwise.
+ * @private
+ */
+function selectorMatches(listener, item) {
+  if (listener.path && item.data) {
+    return has(item.data, listener.path);
+  }
+
+  return true;
+}
+
+module.exports = listenerMatch;
